refactor: Clean up reused boiler plate struct code
