+++ lib/structs/field_content.ex
@@ -75,6 +75,7 @@ defmodule ExPass.Structs.FieldContent do
      This field is required. A date or time value must include a time zone.
   """
 
+  use ExPass.Structs.Base
   use TypedStruct
 
   alias ExPass.Utils.Converter
@@ -267,27 +268,5 @@ defmodule ExPass.Structs.FieldContent do
       |> validate(:value, &Validators.validate_required_value(&1, :value))
 
     struct!(__MODULE__, attrs)
-  end
-
-  defp validate(attrs, key, validator) do
-    case validator.(attrs[key]) do
-      :ok ->
-        attrs
-
-      {:error, reason} ->
-        raise ArgumentError, reason
-    end
-  end
-
-  defimpl Jason.Encoder do
-    def encode(field_content, opts) do
-      field_content
-      |> Map.from_struct()
-      |> Enum.reduce(%{}, fn
-        {_k, nil}, acc -> acc
-        {k, v}, acc -> Map.put(acc, Converter.camelize_key(k), v)
-      end)
-      |> Jason.Encode.map(opts)
-    end
   end
 end
